[Abiraman K] Extracts login mutaion gql
